# docs/swagger.yaml
basePath: /
definitions:
...
  handler.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/handler.UserResponse'
    type: object
...
  handler.IdentifyAnimeRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  handler.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  service.AnimeIdentificationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/service.AnimeIdentificationResponse'
        type: array
    type: object
  service.AnimeIdentificationResponse:
    properties:
      anilist_id:
        type: integer
      episode:
        type: integer
      filename:
        type: string
      from:
        type: number
      image:
        type: string
      similarity:
        type: number
      to:
        type: number
      video:
        type: string
    type: object
  service.AnimeListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/service.AnimeResponse'
        type: array
      pagination:
        $ref: '#/definitions/service.Pagination'
    type: object
  service.AnimeResponse:
    properties:
      episodes:
        type: integer
      external_id:
        type: string
      format:
        type: string
      genres:
        items:
          type: string
        type: array
      image_url:
        type: string
      score:
        type: integer
      season:
        type: string
      season_year:
        type: integer
      status:
        type: string
      synopsis:
        type: string
      title:
        $ref: '#/definitions/service.AnimeTitle'
    type: object
  service.AnimeTitle:
    properties:
      english:
        type: string
      native:
        type: string
      romaji:
        type: string
    type: object
  service.Pagination:
    properties:
      has_next_page:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API untuk personal anime library.
  title: Libra - Anime Library API
  version: "1.0"
paths:
  /animes:
    get:
      description: Get a paginated list of popular anime.
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 50
        minimum: 1
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AnimeListResponse'
        "400":
          description: Invalid pagination parameters
          schema:
            additionalProperties: true
            type: object
        "502":
          description: External API error
          schema:
            additionalProperties: true
            type: object
      summary: Browse anime
      tags:
      - Anime
  /animes/{external_id}:
    get:
      description: Get anime details by AniList external ID.
      parameters:
      - description: AniList anime ID
        in: path
        name: external_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid external ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Anime not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: External API error
          schema:
            additionalProperties: true
            type: object
      summary: Get anime detail
      tags:
      - Anime
  /animes/identify:
    post:
      consumes:
      - application/json
      description: Identify an anime scene from an image URL using trace.moe.
      parameters:
      - description: Image URL
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.IdentifyAnimeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AnimeIdentificationListResponse'
        "400":
          description: Invalid image URL
          schema:
            additionalProperties: true
            type: object
        "502":
          description: External API error
          schema:
            additionalProperties: true
            type: object
      summary: Identify anime from image
      tags:
      - Anime
  /animes/search:
    get:
      description: Search anime by title, external ID, season, year, genre, status,
        or format.
      parameters:
      - description: Anime title
        in: query
        name: title
        type: string
      - description: AniList anime ID
        in: query
        name: external_id
        type: string
      - description: Season
        enum:
        - WINTER
        - SPRING
        - SUMMER
        - FALL
        in: query
        name: season
        type: string
      - description: Season year
        in: query
        name: season_year
        type: integer
      - description: Anime genre
        in: query
        name: genre
        type: string
      - description: Anime status
        enum:
        - FINISHED
        - AIRING
        - NOT_YET_RELEASED
        - CANCELLED
        in: query
        name: status
        type: string
      - description: Anime format
        enum:
        - TV
        - TV_SHORT
        - MOVIE
        - SPECIAL
        - OVA
        - ONA
        - MUSIC
        in: query
        name: format
        type: string
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 50
        minimum: 1
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AnimeListResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "502":
          description: External API error
          schema:
            additionalProperties: true
            type: object
      summary: Search anime
      tags:
      - Anime
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return an authentication token.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and return an authentication token.
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Username or email already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"